// Assets-Uncapped/front-portal-develop/src/services/api/loan-operations/models/Account.ts
export interface Account {
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    iban?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    number?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    sortCode?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    swiftBic?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    blz?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    country?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    accountHolderName?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    currency?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    bankName?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    organisationId?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    connectionId?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Account
     */
    deleted?: boolean;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    externalId?: string;
    /**
     * 
     * @type {string}
     * @memberof Account
     */
    mask?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Account
     */
    tan?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Account
     */
    verified?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Account
     */
    personal?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Account
     */
    partnerVerified?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Account
     */
    mandateVerified?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof Account
     */
    verifiedForPayments?: boolean;
}